add code to download data from zenodo
