+++ compile/dissolve.d.ts
@@ -1,0 +1,5 @@
+declare module "geojson-dissolve" {
+  import { Feature, Geometry } from "geojson";
+  const dissolve: (arg: Feature[]) => Geometry;
+  export default dissolve;
+}
