+++ compile/ld.d.ts
@@ -1,0 +1,34 @@
+export type LD = {
+  "@context": string;
+  "@id": string;
+  "@type": string;
+  creator: Creator[];
+  datePublished: string;
+  description: string;
+  distribution: Distribution[];
+  identifier: string;
+  inLanguage: InLanguage;
+  license: string;
+  name: string;
+  url: string;
+  version: string;
+};
+
+export type Creator = {
+  "@id": string;
+  "@type": string;
+  affiliation: string;
+  name: string;
+};
+
+export type Distribution = {
+  "@type": string;
+  contentUrl: string;
+  encodingFormat: string;
+};
+
+export type InLanguage = {
+  "@type": string;
+  alternateName: string;
+  name: string;
+};
