--- data/dissolve.d.ts
@@ -1,5 +1,0 @@
-declare module "geojson-dissolve" {
-  import { Feature, Geometry } from "geojson";
-  const dissolve: (arg: Feature[]) => Geometry;
-  export default dissolve;
-}
