+++ package.json
@@ -4,18 +4,16 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "compile": "npx ts-node --esm --files ./data/compile.ts",
-    "build": "tsc && vite build",
+    "compile": "npx ts-node --esm --files ./compile/index.ts",
+    "build": "yarn compile && tsc && vite build",
     "preview": "vite preview",
     "lint": "eslint . --ext .tsx,.ts --fix && prettier --write \"**/*.{tsx,ts,css,json,yaml,svg,md}\" --no-error-on-unmatched-pattern"
   },
   "dependencies": {
     "comlink": "^4.4.1",
     "d3": "^7.8.5",
-    "geojson-dissolve": "^3.1.0",
     "gsap": "^3.12.2",
     "lodash": "^4.17.21",
-    "papaparse": "^5.4.1",
     "poisson-disk-sampling": "^2.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -25,6 +23,7 @@
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.0",
+    "@types/cli-progress": "^3.11.0",
     "@types/d3": "^7.4.0",
     "@types/geojson": "^7946.0.10",
     "@types/lodash": "^4.14.195",
@@ -35,6 +34,7 @@
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
     "@vitejs/plugin-react": "^4.0.3",
+    "cli-progress": "^3.12.0",
     "eslint": "^8.45.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
@@ -42,6 +42,10 @@
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
+    "geojson-dissolve": "^3.1.0",
+    "node-html-parser": "^6.1.5",
+    "nodejs-file-downloader": "^4.12.1",
+    "papaparse": "^5.4.1",
     "postcss": "^8.4.26",
     "prettier": "^3.0.0",
     "prettier-plugin-css-order": "^1.3.1",
